+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ### Fixed
 
+* Fixed [#694](https://github.com/NeoVintageous/NeoVintageous/issues/694): `<F2>` and `<S-F2>` next and previous bookmark keys wrong way around
+* Fixed [#693](https://github.com/NeoVintageous/NeoVintageous/issues/693): `<C-S-b>` should show build with overlay
 * Fixed [#10](https://github.com/NeoVintageous/NeoVintageous/issues/10): Repeating text changing doesn't work.
 
 ## 1.19.0 - 2020-02-19